Add checks argument to AddDummyVariablesForNAs
The AddDummyVariablesForNAs function will add dummy variable columns to
the a data.frame. It previously removed rows that had all predictors
missing or the outcome variable was missing. This commit allow the
checks argument to only remove those rows when checks is TRUE. When
checks is FALSE the data with appended dummy columns is returned.

Unit tests for both values of checks added.