DS-2986 Duplicate cols handled in Dummy adjustment
Data that has identical structure in the missing data would cause
identical dummy variables to be created. This is in effect an aliased
predictor in the regression method. This commit avoids this by checking
for duplicates and replacing with a single column. An attribute is added
that maps the dummy variable column to all affected original variables.