Move usage of sanitizeSPSSVariableNames
Moved this so that it occurs just after the initial determination of variable names in the three key functions.

Modified the function so that it modifies a character vector instead of the columns of the variables data frame.

Replaced the de-duplication path in merge data sets by case with the sanitization and ensured "renamed.variables" is still retained as an attribute.

Updated tests where renaming is expected to now also expect a warning.