DS-5384 Nicer message when file has bad encoding (#59)
When an input data file has bad encoding, we throw an error with a nicer message asking the user to reexport the data file to the cloud drive. Ideally we would be able to parse such data files but we are currently limited by what's provided by the haven package. Displayr is able to correctly parse and fix these files when exporting.